+++ modules/core/src/main/scala/io/laminext/core/InputChangeEvents.scala
@@ -5,12 +5,19 @@ import org.scalajs.dom.raw.Event
 
 object InputChangeEvents {
 
-  def apply(el: Element): EventStream[Event] =
+  def apply(
+    el: Element,
+    changeStreamTransform: EventStream[Event] => EventStream[Event] = identity
+  ): EventStream[Event] =
     EventStream
       .merge(
-        el.events(onChange),
+        changeStreamTransform(
+          EventStream.merge(
+            el.events(onChange),
+            el.events(onInput)
+          )
+        ),
         el.events(onBlur),
-        el.events(onInput),
         el.events(onPaste),
         el.events(onCut)
       )