+++ modules/core/src/main/scala/io/laminext/core/ops/htmlelement/TextAreaElementOps.scala
@@ -7,8 +7,15 @@ import org.scalajs.dom.raw.Event
 class TextAreaElementOps(el: TextArea) {
 
   @inline def changes: EventStream[Event] = InputChangeEvents(el)
+  @inline def changes(
+    changeStreamTransform: EventStream[Event] => EventStream[Event]
+  ): EventStream[Event] = InputChangeEvents(el, changeStreamTransform)
 
-  @inline def value: Signal[String] =
-    changes.mapTo(el.ref.value).toSignal(el.ref.value)
+  @inline def value: Signal[String] = value(identity)
+
+  def value(
+    changeStreamTransform: EventStream[Event] => EventStream[Event]
+  ): Signal[String] =
+    changes(changeStreamTransform).mapTo(el.ref.value).toSignal(el.ref.value)
 
 }